since all posts and comments need a date, moved date to its own file
